feat: v1.5.0 双合集AI比对系统 + CUDA/音频编码优化
## 核心特性
- **双合集比对**: tools/compare_collections.py — 全量下载→转录→逐P总结→AI横向比对
- **多P支持**: bilibili_downloader.py 新增 get_video_pages() / download_page_audio()
- **7维度评估**: 知识覆盖面/概念溯源/例题梯度/体系连贯性/直觉辅助/讲解密度/考试适配
- **断点续传**: .done 标记 + _progress.json 支持中断恢复

## 优化
- transcriber: CUDA DLL增强预加载(cublas/cudart/cudnn→os.add_dll_directory+ctypes.CDLL)
- audio_utils: av.AudioResampler 替代手动numpy重采样(dtype正确+flush残帧)
- config: WHISPER_COMPUTE_TYPE=float32 (Blackwell RTX 5070 Ti cuBLAS兼容)
- summarizer: 新增max_tokens参数(4096/8192支持长比对prompt)

## 工具脚本
- compare_collections.py — 主编排脚本(dry-run/resume/compare-only模式)
- comparison_prompt.py — 数学视频结构化摘要+7维评分+JSON→Report生成
- run_pipeline.py — 分阶段管线(download/transcribe/summarize/compare)
- test_5p.py — 合集A前5P验证脚本
- verify_pipeline.py — 18项单元测试(编译/解析/报告生成)

## 审查结果
- 编译: 11/11 文件通过
- 测试: 18/18 项通过
- 修复: 补充 tools/__init__.py | saved_good/ 入 .gitignore

+++ config.py
@@ -44,7 +44,7 @@ def _get(key: str, default=""):
 # ========== Whisper 设置 ==========
 WHISPER_MODEL = "large-v3-turbo"
 WHISPER_DEVICE = "cuda"
-WHISPER_COMPUTE_TYPE = "float16"
+WHISPER_COMPUTE_TYPE = "float32"  # Blackwell RTX 5070 Ti 需 float32 规避 cuBLAS 不兼容（12GB GDDR7 可容纳）
 WHISPER_BEAM_SIZE = 3       # 解码搜索宽度，越小越快（3=快速, 5=平衡, 8=最准）
 # 转写模式: "batch" = 下载完再转写 / "stream" = 边下边转写
 WHISPER_MODE = "batch"
